Updated README.md, added a feature which needs testing: encrypting and compiling.

+++ functions/decompile.py
@@ -1,5 +1,4 @@
 import os
-import pathlib
 
 from functions.other import find_all_lua_files
 
@@ -7,8 +6,17 @@ from functions.other import find_all_lua_files
 def decompile(folder):
     lua_files = find_all_lua_files(folder)
     for item in lua_files:
-        print(item)
-        new_name = f"{item}-decompiled.lua"
-        os.system("java -jar tools/unluac.jar " + f"{item}" + " > " + new_name)
-        os.remove(f"{item}")
-        os.rename(new_name, f"{item}")
+        if is_compiled(item):
+            print(item)
+            new_name = f"{item}-decompiled.lua"
+            os.system("java -jar tools/unluac.jar " + f"{item}" + " > " + new_name)
+            os.remove(f"{item}")
+            os.rename(new_name, f"{item}")
+
+def is_compiled(item):
+    try:
+        with open(item, 'r', encoding='utf-8') as f:
+            data = f.read()
+            return False
+    except:
+        return True
